Refactor survey components
